+++ ts/packages/dispatcher/src/internal.ts
@@ -32,7 +32,7 @@ export {
 export { getBuiltinConstructionConfig } from "./utils/config.js";
 export { getBuiltinSchemaNames } from "./translation/agentTranslators.js";
 export { getAssistantSelectionSchemas } from "./translation/unknownSwitcher.js";
-export { getActionSchema } from "./translation/actionSchema.js";
+export { getActionSchema } from "./translation/actionSchemaFileCache.js";
 export { getTestDataFiles } from "./utils/config.js";
 export {
     loadBuiltinTranslatorSchemaConfig,
